[FEAT/#22] 의존성 수정
