+++ app/build.gradle
@@ -60,7 +60,6 @@ dependencies {
     implementation(project(":feature"))
     implementation(project(":domain"))
     implementation(project(":data"))
-    implementation(project(":core-ui"))
 
     implementation 'androidx.core:core-ktx:1.12.0'
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.7.0'
