+++ app/src/main/java/com/sopt/now/compose/MainActivity.kt
@@ -3,14 +3,7 @@ package com.sopt.now.compose
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.ui.Modifier
-import com.sopt.navigation.MainNavigator
-import com.sopt.navigation.MainScreen
-import com.sopt.navigation.rememberMainNavigator
-import com.sopt.ui.ui.theme.NOWSOPTAndroidTheme
+import com.sopt.navigation.MainApp
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -18,15 +11,7 @@ class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            val navController: MainNavigator = rememberMainNavigator()
-            NOWSOPTAndroidTheme {
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    MainScreen(navController)
-                }
-            }
+            MainApp()
         }
     }
 }
