+++ build.gradle
@@ -7,6 +7,7 @@ plugins {
     id 'org.jetbrains.kotlin.android' version '1.9.0' apply false
     id 'org.jetbrains.kotlin.jvm' version '1.9.0' apply false
     id 'com.google.dagger.hilt.android' version '2.48.1' apply false
+    id "com.vanniktech.dependency.graph.generator" version "0.7.0"
 }
 
 ext {
