+++ navigation/src/main/java/com/sopt/navigation/MainScreen.kt
@@ -2,8 +2,11 @@ package com.sopt.navigation
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
@@ -13,7 +16,20 @@ import com.sopt.feature.auth.signup.navigation.signUpGraph
 import com.sopt.feature.main.home.navigation.homeGraph
 import com.sopt.feature.main.mypage.navigation.myPageGraph
 import com.sopt.feature.main.search.navigation.searchGraph
+import com.sopt.ui.ui.theme.NOWSOPTAndroidTheme
 
+@Composable
+fun MainApp(){
+    val navController: MainNavigator = rememberMainNavigator()
+    NOWSOPTAndroidTheme {
+        Surface(
+            modifier = Modifier.fillMaxSize(),
+            color = MaterialTheme.colorScheme.background
+        ) {
+            MainScreen(navController)
+        }
+    }
+}
 @Composable
 fun MainScreen(
     navController: MainNavigator = rememberMainNavigator()
